리파지토리 userid 롱으로 변경

+++ src/main/java/com/kkanggogo/facealbum/login/domain/repository/UserRepository.java
@@ -13,9 +13,9 @@ public interface UserRepository extends JpaRepository<User,Long> {
     User findByUsername(String username);
 
     @Query(value = "SELECT * FROM User WHERE userId= ?1", nativeQuery = true)
-    Optional<User> searchId(int id);
+    Optional<User> searchId(Long id);
 
-    Optional<User> findById(int id);
+    Optional<User> findById(Long id);
 
 
     @Query(value = "SELECT COUNT(id) FROM User")
